requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosmtplib==4.0.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
environs==14.1.1
fastapi==0.115.8
fastapi-cli==0.0.7
fastapi_cors==0.0.6
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.6.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mongoengine==0.29.1
motor==3.7.0
multidict==6.4.3
orjson==3.10.18
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11
PyMuPDF==1.25.5
pyparsing==3.2.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.0
PyMuPDF
motor
easyocr 
pillow
pydub
SpeechRecognition
